backend/requirements.txt
fastapi==0.119.1
uvicorn==0.38.0
pillow==11.3.0
pytesseract==0.3.13
opencv-python==4.12.0.88
imutils==0.5.4
pydantic==2.12.3
python-multipart==0.0.9
openai==2.6.0
python-dotenv==1.1.1
numpy==2.0.2
